APE kind of working. Flipping trick not incorporated yet
